+++ src/providers/wcs/CMakeLists.txt
@@ -14,6 +14,7 @@ SET (WCS_MOC_HDRS
 
 IF (WITH_GUI)
   SET(WCS_SRCS ${WCS_SRCS}
+    qgswcsprovidergui.cpp
     qgswcssourceselect.cpp
   )
   SET(WCS_MOC_HDRS ${WCS_MOC_HDRS}
